refactor(protocol): move activityRows into the package
The dense activity tree's row-model builder imported only activityData and
stableDelegate, both already in the package, and both frontends already read
it - the web through a relative import and mobile-native through a deep path
into the web tree. It moves into the package with its test file, which stays
the oracle unchanged.

Plumbing so the module is reachable by package name: tsconfig.build.json
emits it, index.ts re-exports its four runtime symbols and six types, and the
qualification manifest names them under "." with smoke calls that exercise
foldRowID, buildActivityRows, jobIsFailed and activityDelegateState from a
bare Node consumer. Without the index.ts re-export the runner now fails the
reachability assertion instead of passing in silence.
