+++ cmd/evener-hub/frontend/src/panes/session/chrome/ActivityRowDetail.tsx
@@ -5,6 +5,7 @@
 // except the one output-tail fetch; ActivityTree owns the detailID state and
 // passes the row plus its ticking `now` straight through.
 import { Fragment, type JSX, useEffect, useState } from "react";
+import { type ActivityDelegateRow, type ActivityJobRow, activityDelegateState } from "../../../protocol/activityRows";
 import { connectionStore } from "../../../stores/connection";
 import { threadsStore } from "../../../stores/threads";
 import { parseAnsiLines } from "../../../widgets/codeblock/ansi";
@@ -15,7 +16,6 @@ import { Markdown } from "../../../widgets/markdown";
 import { formatClockTime, splitMandate } from "../transcript/messages/format";
 import { formatQuietAge, quietAnchorMillis } from "./activityFormat";
 import styles from "./activitypanel.module.css";
-import { type ActivityDelegateRow, type ActivityJobRow, activityDelegateState } from "./activityRows";
 
 const CLASS = {
   detailStrip: requireClass(styles.detailStrip, "activitypanel.module.css", "detailStrip"),
