+++ cmd/evener-hub/frontend/src/protocol/index.ts
@@ -26,6 +26,15 @@ export {
 export type { ActivityBranch, ActivityClient, ActivityState } from "./activityList";
 export { ActivityList } from "./activityList";
 export { fenceRootSession, graftContinuationTree } from "./activityMerge";
+export type {
+  ActivityDelegateRow,
+  ActivityDelegateState,
+  ActivityFoldRow,
+  ActivityJobRow,
+  ActivityRow,
+  ActivityRowBase,
+} from "./activityRows";
+export { activityDelegateState, buildActivityRows, foldRowID, jobIsFailed } from "./activityRows";
 export type { AskAnswerItem, AskResolution } from "./askAnswers";
 export { composeAskAnswers } from "./askAnswers";
 export type { AnyNotification, AppwireClientOptions, ConnectionState, TerminalReason } from "./client";
