+++ cmd/evener-hub/frontend/src/protocol/tsconfig.build.json
@@ -20,6 +20,7 @@
     "types.gen.ts",
     "askAnswers.ts",
     "activityData.ts",
+    "activityRows.ts",
     "activityList.ts",
     "activityMerge.ts",
     "itemFailure.ts",
